Add files via upload
This Project predicts whether the patient is diabetic or not using machine learning algorithms.
The dataset is taken from PIMA Indians and has few features along with a dependent variable "OUTCOME".
After training our model we received a good accuracy in few algorithms. In-order to improve the accuracy , f1score and recall score we did hyper-parameter tuning on few algorithms but due to small dataset I was not able to see major changes in the scores and accuracy . 

This project is done for understanding and to enhance my machine learning skills so that I can apply them correctly in my future projects.